Wrap TcpStream in BufWriter/BufReader
This commit wraps the TcpStream in a BufWriter when sending a request
to a server, and wraps the same TcpStream in a BufReader when reading
the response.  The Rust documentation has this to say about buffering:

> It can be excessively inefficient to work directly with a Read
> instance. For example, every call to read on TcpStream results in a
> system call. [1]

I made a simple test, downloading a [Debian 9.4][2] iso without
buffering, and with buffering.  The numbers are shown below.  We can
see an important reduction in user and system time.

Without buffering
=================

    ~/prog/rust/http-client$ time ./target/release/http-client
    code  : 200
    length: 455808504

    real    4m44.382s
    user    0m43.742s
    sys     1m47.852s

With buffering
==============

    ~/prog/rust/http-client$ time ./target/release/http-client
    code  : 200
    length: 455808504

    real    4m22.748s
    user    0m17.772s
    sys     0m1.504s

[1]: https://doc.rust-lang.org/std/io/struct.BufReader.html
[2]: http://debian.mirror.iweb.ca/debian-cd/9.4.0/amd64/iso-cd/debian-9.4.0-amd64-netinst.iso