Fix timeout behaviour, add a test
As described in issue #22, minreq does not handle timeouts as it implies
in the documentation. This commit fixes it by setting the right timeout*
for each byte read, if a timeout is specified.

* To be specific, it sets each read's timeout to:
  `initial_request_time + timeout - current_time`

Note: this commit might cause some slowdown, as set_read_timeout is
called for each byte read. This only applies to requests with a timeout,
however.