Commit to MSRV 1.48 for current major version
In an effort to support folks that use older toolchains add a section to
the README stating that we commit to building with Rust 1.48 for the
subset of features that excludes all the TLS features. Add an additional
github action to test the MSRV. Note explicitly that this MSRV is per
major version and we reserve the right to change the MSRV when we jump
major versions.

Add a per major version Minimum Supported Rust Version (MSRV).