connection: Don't double-buffer TCP stream
UnsecuredStream doesn't need to be wrapped in a BufWriter since
write_all() passes the entire buffer-to-send until the OS has
written it out.[1] And it doesn't need to be wrapped in a BufReader
since HttpStreamBytes (in response.rs) wraps the whole HttpStream
in a BufReader. Both of these cases cause us to use a little
extra memory and copy bytes twice when we only need to copy once.

[1]: https://doc.rust-lang.org/src/std/io/mod.rs.html#1620-1635