connection: try all the resolved addresses
A couple users encountered an issue connecting to a host which i could
not reproduce. Turns out minreq won't try to connect to all resolved
addresses. This can cause issues for instance if users are on a network
which does not support both v4 and v6.

Fix this by trying to connect to all resolved addresses.