Prune native_tls.rs harder
The TlsConnectorBuilder isn't actually used nor exposed to minreq users,
so its defaults ended up being what it always used, so this just removed
a lot of code that was effectively dead code, just not reported as such
because there was an unused builder API and all the code after that took
all the options into account.

Not too nice to hardcode all these options, but they were hardcoded
before anyways, and if we end up exposing some of these options in the
future, I think it'd be better to handle them *then*, separately for
each tls "backend" we have.