Prune almost everything left of the bundled native-tls openssl code
After "inlining" the functions used from the bundled native_tls code
into the openssl_stream::create_secured_stream function, almost
everything left was discarded. And now it's pretty obvious that using
openssl was about as hard as the other tls backends, except that it
wasn't documented as well, I guess. And there's an Android-specific cert
probing bit.

Also removed build.rs because I don't think it was relevant: we use the
"vendored" features of the openssl crate, which apparently "will
generally track the newest OpenSSL release" which is compiled and linked
statically. So we will definitely always have access to the
set_min_proto_version function.